Store nlist head pointer with size_t type.

+++ hoomd/dem/DEM2DForceComputeGPU.cc
@@ -129,7 +129,7 @@ void DEM2DForceComputeGPU<Real, Real2, Real4, Potential>::computeForces(uint64_t
     ArrayHandle<unsigned int> d_nlist(this->m_nlist->getNListArray(),
                                       access_location::device,
                                       access_mode::read);
-    ArrayHandle<unsigned int> d_head_list(this->m_nlist->getHeadList(),
+    ArrayHandle<size_t> d_head_list(this->m_nlist->getHeadList(),
                                           access_location::device,
                                           access_mode::read);
