+++ hoomd/md/PotentialPairDPDThermoGPU.h
@@ -150,7 +150,7 @@ void PotentialPairDPDThermoGPU<evaluator, gpu_cpdf>::computeForces(uint64_t time
     ArrayHandle<unsigned int> d_nlist(this->m_nlist->getNListArray(),
                                       access_location::device,
                                       access_mode::read);
-    ArrayHandle<unsigned int> d_head_list(this->m_nlist->getHeadList(),
+    ArrayHandle<size_t> d_head_list(this->m_nlist->getHeadList(),
                                           access_location::device,
                                           access_mode::read);
