+++ hoomd/md/PotentialPairGPU.cuh
@@ -54,7 +54,7 @@ struct pair_args_t
                 const BoxDim& _box,
                 const unsigned int* _d_n_neigh,
                 const unsigned int* _d_nlist,
-                const unsigned int* _d_head_list,
+                const size_t* _d_head_list,
                 const Scalar* _d_rcutsq,
                 const Scalar* _d_ronsq,
                 const size_t _size_neigh_list,
@@ -85,7 +85,7 @@ struct pair_args_t
     const unsigned int*
         d_n_neigh;               //!< Device array listing the number of neighbors on each particle
     const unsigned int* d_nlist; //!< Device array listing the neighbors of each particle
-    const unsigned int* d_head_list; //!< Head list indexes for accessing d_nlist
+    const size_t* d_head_list; //!< Head list indexes for accessing d_nlist
     const Scalar* d_rcutsq;          //!< Device array listing r_cut squared per particle type pair
     const Scalar* d_ronsq;           //!< Device array listing r_on squared per particle type pair
     const size_t size_neigh_list;    //!< Size of the neighbor list for texture binding
@@ -151,7 +151,7 @@ gpu_compute_pair_forces_shared_kernel(Scalar4* d_force,
                                       const BoxDim box,
                                       const unsigned int* d_n_neigh,
                                       const unsigned int* d_nlist,
-                                      const unsigned int* d_head_list,
+                                      const size_t* d_head_list,
                                       const typename evaluator::param_type* d_params,
                                       const Scalar* d_rcutsq,
                                       const Scalar* d_ronsq,
@@ -241,7 +241,7 @@ gpu_compute_pair_forces_shared_kernel(Scalar4* d_force,
         if (evaluator::needsCharge())
             qi = __ldg(d_charge + idx);
 
-        unsigned int my_head = d_head_list[idx];
+        size_t my_head = d_head_list[idx];
         unsigned int cur_j = 0;
 
         unsigned int next_j(0);
